Commentaire et messages de console et utilisateur

+++ js/cart.js
@@ -16,6 +16,7 @@ fetch(listeProduits)
     })
     ;
 
+
 function resteDuScript(contenuMagasin) {
     //Recharge la page si le localStorage à changé alors que cette page était déja ouverte dans un autre onglet.
     window.addEventListener("storage", function () {
@@ -26,22 +27,28 @@ function resteDuScript(contenuMagasin) {
     let quantiteTotale = 0;
     let prixTotal = 0
     let panier = lireLocalStorage();
-    //Tri du panier sur le nom
-    trierPanier(panier);
+    // console.log("panier = " + panier);
+    if (panier != null) {
+        //Tri du panier sur le nom
+        trierPanier(panier);
+    }
 
-    //Pour chaque element dans le panier
-    // Construction du dom et cumul quantité et prix
-    for (let index = 0; index < panier.length; index++) {
-        const element = panier[index];
-        //Pour chaque article dans contenuMagasin
-        for (let index = 0; index < contenuMagasin.length; index++) {
-            const article = contenuMagasin[index];
-            //Si egaux construire un element du dom et aglomerer à l'existant
-            //Incremente le prix total et le nombre d'articles
-            if (article.name == element.nom) {
-                quantiteTotale += parseInt(element.quantite);
-                prixTotal += article.price * parseInt(element.quantite);
-                contenuPanierDom += constructionElementDOM(element, article);
+    // Si un panier existe dans le localStorage
+    if (panier != null) {
+        //Pour chaque element dans le panier
+        // Construction du dom et cumul quantité et prix
+        for (let index = 0; index < panier.length; index++) {
+            const element = panier[index];
+            //Pour chaque article dans contenuMagasin
+            for (let index = 0; index < contenuMagasin.length; index++) {
+                const article = contenuMagasin[index];
+                //Si egaux construire un element du dom et aglomerer à l'existant
+                //Incremente le prix total et le nombre d'articles
+                if (article.name == element.nom) {
+                    quantiteTotale += parseInt(element.quantite);
+                    prixTotal += article.price * parseInt(element.quantite);
+                    contenuPanierDom += constructionElementDOM(element, article);
+                }
             }
         }
     }
@@ -53,46 +60,51 @@ function resteDuScript(contenuMagasin) {
     majQuantiteDom(quantiteTotale);
     majPrixDom(prixTotal);
 
-    //Pour tout les articles dans la page
-    //Ajoute un ecouteur sur un changement de quantite qui modifie le panier et le localStorage en fonction
-    for (let index = 0; index < panier.length; index++) {
-        cibleDom.children[index].querySelector(".itemQuantity").addEventListener("change", function () {
-            let cibleDom = document.getElementById("cart__items");
-            let id = cibleDom.children[index].dataset.id;
-            let couleur = cibleDom.children[index].dataset.color;
-            let quantite = cibleDom.children[index].querySelector(".itemQuantity").value;
-            cibleDom.children[index].querySelector(".itemQuantity").value = quantite;
-
-            if (quantite > 0 & quantite <= 100 & quantite != null & quantite != NaN & quantite != "") {
-                quantite = parseInt(quantite);
-                //Met à jour la quantité totale sur la page                
+    // Si un panier existe dans le localStorage
+    if (panier != null) {
+        //Pour tout les articles dans la page
+        //Ajoute un ecouteur sur un changement de quantite qui modifie le panier et le localStorage en fonction
+        for (let index = 0; index < panier.length; index++) {
+            cibleDom.children[index].querySelector(".itemQuantity").addEventListener("change", function () {
+                let cibleDom = document.getElementById("cart__items");
+                let id = cibleDom.children[index].dataset.id;
+                let couleur = cibleDom.children[index].dataset.color;
+                let quantite = cibleDom.children[index].querySelector(".itemQuantity").value;
                 cibleDom.children[index].querySelector(".itemQuantity").value = quantite;
-                // De combien la quantité à changé ?
-                let quantitePanier = panier[index].quantite;
-                let differenceQuantite = quantitePanier - quantite;
-                // Met à jour
-                quantiteTotale -= differenceQuantite;
-                majQuantiteDom(quantiteTotale);
-                // Quel est le prix de l'article en cours de traitement
-                let prixArticle = trouvePrixArticle(contenuMagasin, id);
-                //Met à jour le prix total et le total sur la page
-                prixTotal -= differenceQuantite * prixArticle;
-                majPrixDom(prixTotal);
-                //Modifie la quantité dans le localStorage
-                modifierQuantité(panier, id, couleur, quantite);
-            } else {
-                // L'entrée n'est pas valide, reaffiche le contenu du panier
-                cibleDom.children[index].querySelector(".itemQuantity").value = panier[index].quantite;
-            }
-        });
-        //Ajoute un ecouteur sur la suppression qui modifie le localStorage en fonction et recharge la page
-        cibleDom.children[index].querySelector(".deleteItem").addEventListener("click", function () {
-            let cibleDom = document.getElementById("cart__items");
-            let id = cibleDom.children[index].dataset.id;
-            let couleur = cibleDom.children[index].dataset.color;
-            retirerLocalStorage(panier, id, couleur);
-        });
+
+                if (quantite > 0 & quantite <= 100 & quantite != null & quantite != NaN & quantite != "") {
+                    quantite = parseInt(quantite);
+                    //Met à jour la quantité totale sur la page                
+                    cibleDom.children[index].querySelector(".itemQuantity").value = quantite;
+                    // De combien la quantité à changé ?
+                    let quantitePanier = panier[index].quantite;
+                    let differenceQuantite = quantitePanier - quantite;
+                    // Met à jour
+                    quantiteTotale -= differenceQuantite;
+                    majQuantiteDom(quantiteTotale);
+                    // Quel est le prix de l'article en cours de traitement
+                    let prixArticle = trouvePrixArticle(contenuMagasin, id);
+                    //Met à jour le prix total et le total sur la page
+                    prixTotal -= differenceQuantite * prixArticle;
+                    majPrixDom(prixTotal);
+                    //Modifie la quantité dans le localStorage
+                    modifierQuantité(panier, id, couleur, quantite);
+                } else {
+                    // L'entrée n'est pas valide, reaffiche le contenu du panier
+                    cibleDom.children[index].querySelector(".itemQuantity").value = panier[index].quantite;
+                }
+            });
+            //Ajoute un ecouteur sur la suppression qui modifie le localStorage en fonction et recharge la page
+            cibleDom.children[index].querySelector(".deleteItem").addEventListener("click", function () {
+                let cibleDom = document.getElementById("cart__items");
+                let id = cibleDom.children[index].dataset.id;
+                let couleur = cibleDom.children[index].dataset.color;
+                retirerLocalStorage(panier, id, couleur);
+            });
+        }
+
     }
+
 
     // Debut du traitement du formulaire**************************
 
@@ -106,10 +118,10 @@ function resteDuScript(contenuMagasin) {
     }
 
     // Si localStorage contact n'existe pas, il est créé avec des valeurs vides.
-    if (localStorage.getItem("contact") == null){
-        localStorage.setItem("contact",JSON.stringify(contact));
+    if (localStorage.getItem("contact") == null) {
+        localStorage.setItem("contact", JSON.stringify(contact));
     }
-    // Si il existe ou pas, lecture des valeurs enregistrées.
+    // Lecture des valeurs de contact enregistrées.
     let contactStorage = JSON.parse(localStorage.getItem("contact"));
     // Pour tout les input
     // Si une valeur valide est presente dans le localStorage
@@ -168,7 +180,7 @@ function resteDuScript(contenuMagasin) {
         // Prerempli l'input en cours par la derniere donnée valide connue
         // Et bascule le drapeau valide dans la table de verification à vrai
         cible.value = contact[element.label];
-        if (cible.value != ""){
+        if (cible.value != "") {
             element.valide = true
         }
         // Place un ecouteur sur changement
@@ -178,12 +190,12 @@ function resteDuScript(contenuMagasin) {
             console.log(regExVal);
             // Input valide ?
             if (cible.value.match(regExVal)) {
-                console.log("Entrée valide");
+                console.log("Entrée valide " + element.label);
                 // Bascule le drapeau valide dans la table de verification à vrai
                 element.valide = true;
                 // Met à jour contact et contactStorage avec la nouvelle donnée valide              
                 contact[element.label] = cible.value;
-                majContactStorage(contact);               
+                majContactStorage(contact);
             } else {
                 console.log("Entrée non autorisée!");
                 // Efface l'entrée erronée 
@@ -202,51 +214,58 @@ function resteDuScript(contenuMagasin) {
         });
     });
 
-    // Ecoute le bouton commander et reagi au click
-    document.getElementById("order").addEventListener("click", function (e) {
-        e.preventDefault();
-        // Verifie que le panier n'est pas vide
-        if (panier.length != 0) {
-            // Indique bon pour commande
-            let flagCommander = true;
-            // Pour chaque input du formulaire
-            tableVerification.forEach(element => {
-                console.log(`Imput du formulaire bien rempli ? ${element.valide}`);
-                // Verifie si valide
-                if (element.valide == false) {
-                    // Si invalide, indicateur bon pour commande = false
-                    flagCommander = false;
+    // Si un panier existe dans le localStorage
+    if (panier != null) {
+        // Ecoute le bouton commander et reagi au click
+        document.getElementById("order").addEventListener("click", function (e) {
+            e.preventDefault();
+            // Verifie que le panier n'est pas vide
+            if (panier.length != 0) {
+                // Indique bon pour commande
+                let flagCommander = true;
+                // Pour chaque input du formulaire
+                tableVerification.forEach(element => {
+                    console.log(`Imput du formulaire bien rempli ? ${element.valide}`);
+                    // Verifie si valide
+                    if (element.valide == false) {
+                        // Si invalide, indicateur bon pour commande = false
+                        flagCommander = false;
+                    }
+                });
+                // Si tout les input sont valides
+                if (flagCommander == true) {
+                    // alert("Passer la commande");
+                    location.replace("confirmation.html");
+                } else {
+                    // Un au moins des input est invalide
+                    alert("Le formulaire n'est pas bien rempli!");
                 }
-            });
-            // Si tout les input sont valides
-            if (flagCommander == true) {
-                alert("Passer la commande");
-                location.replace("confirmation.html");
+                console.log(`Le panier contient ${panier.length} article`);
+
             } else {
-                // Un au moins des input est invalide
-                alert("Le formulaire n'est pas bien rempli!");
+
             }
-            console.log(`Le panier contient ${panier.length} article`);
+        })
+    } else {
+        alert("Panier inexistant !");
+    }
+    console.log("Attente action utilisateur");
 
-        } else {
-            alert("Le panier est vide !")
-        }
-
-    })
-
-}
+}// ************** Fin ***********
 
 //Definitions de fonctions***************************************
 
 
 // Met à jour les infos de contact dans le localStorage
 function majContactStorage(contact) {
+    console.log("Fonction mise à jour contact dans le localStorage");
     let contactJson = JSON.stringify(contact);
     localStorage.setItem("contact", contactJson);
 }
 
 //Parcours le tableau des articles du serveur et retourne le prix correspondant à l'id passé en parametre
 function trouvePrixArticle(contenuMagasin, id) {
+    console.log("Fonction trouve le prix d'un article");
     for (let index = 0; index < contenuMagasin.length; index++) {
         const element = contenuMagasin[index];
         if (element._id == id) {
@@ -257,12 +276,14 @@ function trouvePrixArticle(contenuMagasin, id) {
 
 // Met à jour le prix sur la page
 function majPrixDom(prixTotal) {
+    console.log("Fonction met le prix à jour sur la page");
     cibleDom = document.getElementById("totalPrice");
     cibleDom.innerHTML = prixTotal;
 }
 
 // Met à jour la quantité sur la page
 function majQuantiteDom(quantiteTotale) {
+    console.log("fonction met la quantité à jour sur la page");
     cibleDom = document.getElementById("totalQuantity");
     cibleDom.innerHTML = quantiteTotale;
 
@@ -270,6 +291,7 @@ function majQuantiteDom(quantiteTotale) {
 
 // Retire un article du panier et du localStorage puis recharge la page
 function retirerLocalStorage(panier, id, couleur) {
+    console.log("Fonction retire un article du localStorage");
     for (let index = 0; index < panier.length; index++) {
         const element = panier[index];
         if (element.id == id & element.couleur == couleur) {
@@ -287,6 +309,7 @@ function retirerLocalStorage(panier, id, couleur) {
 
 // Modifie la quantité d'un article dans le panier et le localStorage
 function modifierQuantité(panier, id, couleur, quantite) {
+    console.log("Fonction modifier quantité dans le localStorage");
     panier.forEach(element => {
         if (element.id == id & element.couleur == couleur) {
             element.quantite = quantite;
@@ -297,14 +320,15 @@ function modifierQuantité(panier, id, couleur, quantite) {
 
 // Met à jour le localStorage avec le contenu du panier
 function majLocalStorage(panier) {
+    console.log("Fonction mise à jour du localStorage");
     let panierJson = JSON.stringify(panier);
     localStorage.setItem("panierKanap", panierJson);
-    console.log("Panier mis à jour");
 }
 
 // Trie le panier par ordre alphabetique
 // Permet de regrouper les articles de meme id et de couleurs differentes
 function trierPanier(panier) {
+    console.log("Fonction trier le panier");
     panier.sort(function compare(a, b) {
         if (a.nom < b.nom)
             return -1;
@@ -316,6 +340,7 @@ function trierPanier(panier) {
 
 // Construit un élément du Dom
 function constructionElementDOM(element, article) {
+    console.log("Fonction construction d'un élément du Dom");
     contenuUnElement = `
     <article class="cart__item" data-id="${element.id}" data-color="${element.couleur}">\n
     <div class="cart__item__img">\n
@@ -344,5 +369,10 @@ function constructionElementDOM(element, article) {
 
 // Lecture du panier sur le localStorage
 function lireLocalStorage() {
-    return JSON.parse(localStorage.getItem("panierKanap"));
+    console.log("Fonction lire le localStorage");
+    if (localStorage.getItem("panierKanap") != null) {
+        return JSON.parse(localStorage.getItem("panierKanap"));
+    } else {
+        return null;
+    }
 }
